pyproject.toml
[project]
name = "youtube-downloader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "yt-dlp>=2025.8.11",
    "python-docx>=1.1.2",
    "python-pptx>=0.6.23",
    "PyPDF2>=3.0.1",
    "markdown>=3.6",
    "pypandoc>=1.13",
    "pillow>=10.4.0",
    "openpyxl>=3.1.5",
]
